+++ src/background/state/application/plugins.rs
@@ -28,8 +28,9 @@ impl FullState {
             }
             match Plugin::load(&path) {
                 Ok(mut plugin) => {
-                    plugin.metadata.bundled = path.starts_with(bundled_path);
-                    self.plugins.insert(plugin.metadata.path.clone(), plugin);
+                    plugin.metadata.internal.bundled = path.starts_with(bundled_path);
+                    self.plugins
+                        .insert(plugin.metadata.internal.path.clone(), plugin);
                 }
                 Err(e) => {
                     log::error!("Failed to load plugin: {e}");